requirements.txt
PyQt5>=5.15.0
pyautogui==0.9.53
pywin32>=306
pillow>=8.0.0
uiautomation==2.0.17
pandas>=1.3.0
openpyxl>=3.0.0
keyboard>=0.13.5
pyinstaller>=5.6.2
pywin32-ctypes>=0.2.0 